validaciones
Validaciones en productos

+++ msv-productos/src/main/java/com/ecommerce/productos/services/ProductoServiceImpl.java
@@ -2,24 +2,33 @@ package com.ecommerce.productos.services;
 
 import java.util.List;
 
+import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.ecommerce.commons.dto.ProductoRequest;
 import com.ecommerce.commons.dto.ProductoResponse;
+import com.ecommerce.commons.exceptions.EntidadRelacionadaException;
+import com.ecommerce.productos.clients.PedidoClient;
 import com.ecommerce.productos.entities.Producto;
 import com.ecommerce.productos.mappers.ProductoMapper;
 import com.ecommerce.productos.repositories.ProductoRepository;
 
-public class ProductoServiceImpl implements ProductoService{
+@Service
+public class ProductoServiceImpl  implements ProductoService{
 
-	private final ProductoRepository repository;
+    private final ProductoRepository repository;
 	
 	private final ProductoMapper mapper;
+	
+	private final PedidoClient pedidoClient;
 
-	public ProductoServiceImpl(ProductoRepository repository, ProductoMapper mapper) {
+	
+	
+	public ProductoServiceImpl(ProductoRepository repository, ProductoMapper mapper, PedidoClient pedidoClient) {
 		super();
 		this.repository = repository;
 		this.mapper = mapper;
+		this.pedidoClient = pedidoClient;
 	}
 
 	@Override
@@ -57,9 +66,16 @@ public class ProductoServiceImpl implements ProductoService{
 	@Transactional
 	public ProductoResponse eliminar(Long id) {
 		Producto producto = repository.findById(id).orElseThrow();
+		
+		boolean isPresent = pedidoClient.productoIsPresent(id);
+		if (isPresent) {
+			throw new EntidadRelacionadaException("No puede eliminar el producto porque esta en un pedido pendiente");
+			
+		}
 		repository.deleteById(id);
 		return mapper.entityToResponse(producto);
 	}
+
